set up bing image search api
