+++ package.json
@@ -8,7 +8,8 @@
     "start": "node server.js",
     "lint": "eslint --quiet .",
     "fix": "eslint --fix .",
-    "watch": "nodemon server.js"
+    "watch": "nodemon server.js",
+    "bing": "nodemon ./routes/bing.js"
   },
   "keywords": [],
   "author": "",
@@ -32,6 +33,7 @@
     "eslint-config-prettier": "^2.9.0",
     "eslint-plugin-prettier": "^2.6.2",
     "mocha": "^5.2.0",
+    "nodemon": "^2.0.6",
     "prettier": "^1.13.7"
   }
 }
